--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -1,8 +1,11 @@
-module github.com/fixme_my_friend/hw10_program_optimization
+module github.com/zoolberc/otus-hw/hw10_program_optimization
 
 go 1.19
 
-require github.com/stretchr/testify v1.7.0
+require (
+	github.com/stretchr/testify v1.7.0
+	github.com/valyala/fastjson v1.6.4
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect