### Критерии оценки
- Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- Добавлены юнит-тесты - до 3 баллов
- Понятность и чистота кода - до 3 баллов

#### Old result
RUN   TestGetDomainStat_Time_And_Memory
    stats_optimization_test.go:46: time used: 350.2007ms / 300ms
    stats_optimization_test.go:47: memory used: 308Mb / 30Mb
    assertion_compare.go:332: 
        	Error Trace:	stats_optimization_test.go:49
        	Error:      	"350200700" is not less than "300000000"
        	Test:       	TestGetDomainStat_Time_And_Memory
        	Messages:   	[the program is too slow]
        	
#### New result 
=== RUN   TestGetDomainStat_Time_And_Memory
    stats_optimization_test.go:46: time used: 92.2167ms / 300ms
    stats_optimization_test.go:47: memory used: 15Mb / 30Mb
--- PASS: TestGetDomainStat_Time_And_Memory (1.03s)

#### Bench test
goos: windows
goarch: amd64
pkg: github.com/zoolberc/otus-hw/hw10_program_optimization
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
BenchmarkGetDomainStat-20    	 1549795	       718.4 ns/op	    4154 B/op	       2 allocs/op
BenchmarkGetDomainStat-20    	 1521585	       757.1 ns/op	    4154 B/op	       2 allocs/op
BenchmarkGetDomainStat-20    	 1518202	       739.9 ns/op	    4154 B/op	       2 allocs/op
BenchmarkGetDomainStat-20    	 1404535	       793.4 ns/op	    4155 B/op	       2 allocs/op
BenchmarkGetDomainStat-20    	 1348939	       811.0 ns/op	    4156 B/op	       2 allocs/op
PASS
ok  	github.com/zoolberc/otus-hw/hw10_program_optimization	13.274s